Example CLI utility mvs-render
