--- examples/1cbs.mvsj
@@ -1,46 +1,0 @@
-{
- "version": 6,
- "root": {
-  "kind": "root",
-  "children": [
-   {
-    "kind": "download",
-    "params": {
-     "url": "http://localhost:8080/examples/1cbs_updated.cif"
-    },
-    "children": [
-     {
-      "kind": "parse",
-      "params": {
-       "format": "mmcif"
-      },
-      "children": [
-       {
-        "kind": "structure",
-        "params": {
-         "type": "model"
-        },
-        "children": [
-         {
-          "kind": "component",
-          "params": {
-           "selector": "all"
-          },
-          "children": [
-           {
-            "kind": "representation",
-            "params": {
-             "type": "cartoon"
-            }
-           }
-          ]
-         }
-        ]
-       }
-      ]
-     }
-    ]
-   }
-  ]
- }
-}
