+++ src/extensions/mvs/load-helpers.ts
@@ -59,13 +59,7 @@ export async function loadTree<TTree extends Tree, TContext>(plugin: PluginConte
             }
         }
     });
-    console.log('commiting...')
-    try {
-        await update.commit();
-    } catch (ex) {
-        console.log('caught')
-    }
-    console.log('DONE commit')
+    await update.commit();
 }
